cyclades: Fix bug with volume quotas
When computing 'cyclades.disk' resource for a 'DESTROY' server action,
count only the volumes that are in the 'IN_USE' status. If the
volumes are not in this status, then a commission has already been
issued for that action. However, there is an exception, if the server is
in 'BUILD' operstate, where we must also count the 'CREATING' volumes.