Router gets the gateway ip of the subnets it holds
Gateway IP is not in the pool of addresses that ganeti can allocate
but in the case of the router it is passed using a no conflicts check.
Since a user can only have one router and noone else can use the gw ip
of a private subnet it is safe to assign it with the no-conflicts
option.